refactor: remove BypassStateManager indirection, clean up executeAgent
- Delete BypassStateManager class and inline the simple Map<StreamTabId, boolean>
  pattern directly into toolEditApproval.ts and proposalApproval.ts. The class
  added generic wrappers and unsafe casts (`as never`) around a trivial data
  structure used by only two consumers.
- Remove bare block scope in executeAgent that was introduced to limit
  `runStorage` visibility; inline the single call instead.
